+++ Tests/MapboxMapsTests/Gestures/GestureHandlers/SingleTapGestureHandlerTests.swift
@@ -1,0 +1,48 @@
+import XCTest
+@testable import MapboxMaps
+
+final class SingleTapGestureHandlerTests: XCTestCase {
+
+    var gestureRecognizer: MockTapGestureRecognizer!
+    var cameraAnimationsManager: MockCameraAnimationsManager!
+    var mapboxMap: MockMapboxMap!
+    var gestureHandler: SingleTapGestureHandler!
+    // swiftlint:disable:next weak_delegate
+    var delegate: MockGestureHandlerDelegate!
+
+    override func setUp() {
+        super.setUp()
+        gestureRecognizer = MockTapGestureRecognizer()
+        cameraAnimationsManager = MockCameraAnimationsManager()
+        mapboxMap = MockMapboxMap()
+        gestureHandler = SingleTapGestureHandler(
+            gestureRecognizer: gestureRecognizer,
+            mapboxMap: mapboxMap,
+            cameraAnimationsManager: cameraAnimationsManager)
+        delegate = MockGestureHandlerDelegate()
+        gestureHandler.delegate = delegate
+    }
+
+    override func tearDown() {
+        delegate = nil
+        gestureHandler = nil
+        mapboxMap = nil
+        cameraAnimationsManager = nil
+        gestureRecognizer = nil
+        super.tearDown()
+    }
+
+    func testInitialization() {
+        XCTAssertTrue(gestureRecognizer === gestureHandler.gestureRecognizer)
+        XCTAssertEqual(gestureRecognizer.numberOfTapsRequired, 1)
+        XCTAssertEqual(gestureRecognizer.numberOfTouchesRequired, 1)
+    }
+
+    func testHandler() {
+        gestureRecognizer.getStateStub.defaultReturnValue = .recognized
+        gestureRecognizer.sendActions()
+
+        XCTAssertEqual(delegate.gestureBeganStub.parameters, [.singleTap])
+        XCTAssertEqual(cameraAnimationsManager.cancelAnimationsStub.invocations.count, 1)
+    }
+}